Decouple AdminBundle from the Core - related to https://github.com/pimcore/pimcore/issues/12531
