+++ bundles/AdminBundle/Resources/config/aliases.yaml
@@ -3,20 +3,20 @@ services:
         public: true
 
     pimcore_admin.security.user_provider:
-        alias: Pimcore\Bundle\AdminBundle\Security\User\UserProvider
+        alias: Pimcore\Security\User\UserProvider
         deprecated:
             package: 'pimcore/pimcore'
             version: '10.6'
             message: 'The "%alias_id%" alias is deprecated and will be removed in Pimcore 11'
-    pimcore_admin.security.token_storage_user_resolver: '@Pimcore\Bundle\AdminBundle\Security\User\TokenStorageUserResolver'
+    pimcore_admin.security.token_storage_user_resolver: '@Pimcore\Security\User\TokenStorageUserResolver'
     pimcore_admin.security.user_loader:
-        alias: Pimcore\Bundle\AdminBundle\Security\User\UserLoader
+        alias: Pimcore\Security\User\UserLoader
         deprecated:
             package: 'pimcore/pimcore'
             version: '10.6'
             message: 'The "%alias_id%" alias is deprecated and will be removed in Pimcore 11'
     pimcore_admin.security.user_checker:
-        alias: Pimcore\Bundle\AdminBundle\Security\User\UserChecker
+        alias: Pimcore\Security\User\UserChecker
         deprecated:
             package: 'pimcore/pimcore'
             version: '10.6'
