+++ bundles/AdminBundle/Resources/views/Admin/Index/index.html.twig
@@ -1,5 +1,5 @@
 {% set language = app.request.locale %}
-{# @var  \Pimcore\Bundle\AdminBundle\Security\User\User #}
+{# @var  \Pimcore\Security\User\User #}
 {% set userProxy = app.user %}
 {% set user = userProxy.user %}
