+++ bundles/AdminBundle/Security/Authenticator/AdminLoginAuthenticator.php
@@ -15,7 +15,7 @@
 
 namespace Pimcore\Bundle\AdminBundle\Security\Authenticator;
 
-use Pimcore\Bundle\AdminBundle\Security\User\User;
+use Pimcore\Security\User\User;
 use Pimcore\Event\Admin\Login\LoginFailedEvent;
 use Pimcore\Event\Admin\Login\LoginRedirectEvent;
 use Pimcore\Event\AdminEvents;
