+++ bundles/CoreBundle/Resources/config/authenticator_security.yaml
@@ -3,7 +3,7 @@ pimcore:
     security:
         factory_type: password_hasher
         password_hasher_factories:
-            Pimcore\Bundle\AdminBundle\Security\User\User: pimcore_admin.security.password_hasher_factory
+            Pimcore\Security\User\User: pimcore_admin.security.password_hasher_factory
 
 scheb_two_factor:
     security_tokens:
