+++ doc/Development_Documentation/20_Extending_Pimcore/24_Add_Your_Own_Permissions.md
@@ -24,7 +24,7 @@ class AdminController extends PimcoreAdminController
      */
     public function myAdminAction(Request $request) {
 
-        /** @var \Pimcore\Bundle\AdminBundle\Security\User\User $user */
+        /** @var \Pimcore\Security\User\User $user */
         $user = $this->getUser();
         $pimcoreUser = $user->getUser();
