+++ lib/Extension/Bundle/PimcoreBundleManager.php
@@ -17,7 +17,6 @@ declare(strict_types=1);
 
 namespace Pimcore\Extension\Bundle;
 
-use Pimcore\Bundle\AdminBundle\Support\PimcoreBundleAdminSupportInterface;
 use Pimcore\Event\BundleManager\PathsEvent;
 use Pimcore\Event\BundleManagerEvents;
 use Pimcore\Extension\Bundle\Config\StateConfig;
@@ -674,7 +673,7 @@ class PimcoreBundleManager
         foreach ($this->getActiveBundles() as $bundle) {
             $paths = $bundle->$getter();
 
-            if (!empty($paths) && !$bundle instanceof PimcoreBundleAdminSupportInterface) {
+            if (!empty($paths) && !$bundle instanceof Pimcore) {
                 trigger_deprecation(
                     'pimcore/pimcore',
                     '10.6',
@@ -682,7 +681,7 @@ class PimcoreBundleManager
                         'Calling %s::%s without implementing the %s is deprecated and won\'t be possible in Pimcore 11.',
                         get_class($bundle),
                         $getter,
-                        PimcoreBundleAdminSupportInterface::class
+                        PimcoreBundleAdminClassicInterface::class
                     )
                 );
             }
