+++ lib/Http/Request/Resolver/EditmodeResolver.php
@@ -15,7 +15,7 @@
 
 namespace Pimcore\Http\Request\Resolver;
 
-use Pimcore\Bundle\AdminBundle\Security\User\UserLoader;
+use Pimcore\Security\User\UserLoader;
 use Pimcore\Http\RequestHelper;
 use Psr\Log\LoggerAwareInterface;
 use Psr\Log\LoggerAwareTrait;
