+++ lib/Security/User/UserProvider.php
@@ -13,7 +13,7 @@
  *  @license    http://www.pimcore.org/license     GPLv3 and PCL
  */
 
-namespace Pimcore\Bundle\AdminBundle\Security\User;
+namespace Pimcore\Security\User;
 
 use Pimcore\Model\User as PimcoreUser;
 use Symfony\Component\Security\Core\Exception\UnsupportedUserException;
